添加了用户详情界面 点击弹幕查id结果可以跳转至用户信息界面

+++ v3/app/src/main/java/com/meng/biliv3/fragment/LiveFragment.java
@@ -27,7 +27,7 @@ public class LiveFragment extends BaseIdFragment {
 	private Spinner selectAccount;
 
 	private JsonObject liveInfo;
-	public LiveFragment(String type, int liveId) {
+	public LiveFragment(String type, long liveId) {
 		this.type = type;
 		id = liveId;
 	}
